benchmark/requirements.txt
polytope-client>=0.7.8
pyproj>=3.6.0
pyyaml>=6.0
